Move Flatten and WithFixedSize to interfaces to improve discoverability.
This fixes https://github.com/GoogleCloudPlatform/gcloud-dotnet/issues/157
