+++ src/Google.Api.Gax/PageStreamingInterfaces.cs
@@ -18,6 +18,28 @@ namespace Google.Api.Gax
 
     public interface IPagedAsyncEnumerable<TResponse, TResource> : IPagedAsyncEnumerable<TResponse>
     {
+        /// <summary>
+        /// Flattens this asynchronous sequence of "pages of resources" to an asynchronous sequence of resources.
+        /// </summary>
+        /// <remarks>
+        /// This is expected to be equivalent to calling <c>SelectMany(page => page.ToAsyncEnumerable())</c>, and could
+        /// be implemented as an extension method, but is part of the interface to provide greater discoverability.
+        /// </remarks>
+        /// <returns>A sequence of resources.</returns>
+        IAsyncEnumerable<TResource> Flatten();
+
+        /// <summary>
+        /// Returns an asynchronous sequence of fixed-size pages, so that only the final page returned can have fewer items than
+        /// <paramref name="pageSize"/>. API requests are made appropriately to ensure that the page token returned at
+        /// the end of each page reflects exactly that page boundary; this may entail more API requests than iterating
+        /// over the sequence without enforcing fixed page sizes.
+        /// </summary>
+        /// <param name="pageSize">The size of the pages to return. Must be positive.</param>
+        /// <remarks>
+        /// This could be implemented as an extension method, but is part of the interface to provide greater discoverability.
+        /// </remarks>
+        /// <returns>A sequence of fixed-size pages.</returns>
+        IAsyncEnumerable<FixedSizePage<TResource>> WithFixedPageSize(int pageSize);
     }
 
     public interface IPagedAsyncEnumerator<T> : IAsyncEnumerator<T>
@@ -32,6 +54,28 @@ namespace Google.Api.Gax
 
     public interface IPagedEnumerable<TResponse, TResource> : IPagedEnumerable<TResponse>
     {
+        /// <summary>
+        /// Flattens this sequence of "pages of resources" to a sequence of resources.
+        /// </summary>
+        /// <remarks>
+        /// This is expected to be equivalent to calling <c>SelectMany(page => page)</c>, and could
+        /// be implemented as an extension method, but is part of the interface to provide greater discoverability.
+        /// </remarks>
+        /// <returns>A sequence of resources.</returns>
+        IEnumerable<TResource> Flatten();
+
+        /// <summary>
+        /// Returns a sequence of fixed-size pages, so that only the final page returned can have fewer items than
+        /// <paramref name="pageSize"/>. API requests are made appropriately to ensure that the page token returned at
+        /// the end of each page reflects exactly that page boundary; this may entail more API requests than iterating
+        /// over the sequence without enforcing fixed page sizes.
+        /// </summary>
+        /// <param name="pageSize">The size of the pages to return. Must be positive.</param>
+        /// <remarks>
+        /// This could be implemented as an extension method, but is part of the interface to provide greater discoverability.
+        /// </remarks>
+        /// <returns>A sequence of fixed-size pages.</returns>
+        IEnumerable<FixedSizePage<TResource>> WithFixedPageSize(int pageSize);
     }
 
     public interface IPagedEnumerator<T> : IEnumerator<T>
